+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Component] What are the breaking changes?
 - [Tag] Rewritten Tag component
+- [Button] Renewed Button component with theming support
+- [Accordion] Changed size prop to use AccordionSize enum (AccordionSize.Small, AccordionSize.Medium and AccordionSize.Large) or without usin Typescript ("small", "medium" and "large"). Also theming has changed variables.
 
 #### Added
 
@@ -34,6 +36,7 @@ Changes that are not related to specific components
 
 - [Component] What are the breaking changes?
 - [Tag] Rewritten Tag component
+- [Button] Renewed Button component with theming support
 
 #### Added
 
@@ -65,6 +68,7 @@ Changes that are not related to specific components
 
 - [Component] What has been changed
 - [Tag] Document the new Tag component and it's uses
+- [Button] Renewed Button component with theming support
 
 #### Fixed